% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skrapning.R
\name{hitta_alla}
\alias{hitta_alla}
\title{Hitta alla matchande element i en skrapsession (genväg)}
\usage{
hitta_alla(skrap, ...)
}
\arguments{
\item{skrap}{Ett objekt skapat av starta_skrapsession().}

\item{...}{Vidare till selenider::find_elements().}
}
\description{
Som hitta(), men motsvarar selenider::find_elements()/ss().
}
