% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspelning.R
\name{lasa_av_inspelning}
\alias{lasa_av_inspelning}
\title{Läs av alla nya inspelade händelser och töm arrayen i webbläsaren}
\usage{
lasa_av_inspelning(skrap)
}
\arguments{
\item{skrap}{Ett objekt skapat av starta_skrapsession().}
}
\value{
En tibble, en rad per händelse (kan ha 0 rader).
}
\description{
Läs av alla nya inspelade händelser och töm arrayen i webbläsaren
}
