% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skrapning.R
\name{with_skrapsession}
\alias{with_skrapsession}
\title{Kör kod med en skrapsession och städa bort den automatiskt efteråt}
\usage{
with_skrapsession(expr, ...)
}
\arguments{
\item{expr}{Kod att köra. En variabel med namnet \code{skrap} finns
tillgänglig i uttrycket och pekar på sessionsobjektet.}

\item{...}{Argument som skickas vidare till starta_skrapsession().}
}
\description{
Bekvämlighetsfunktion som garanterar städning även om koden i uttrycket
kastar ett fel.
}
\examples{
\dontrun{
with_skrapsession({
  selenider::open_url(skrap$session, "https://www.regiondalarna.se")
  selenider::s(skrap$session, "title") |> selenider::elem_text()
})
}
}
