% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspelning.R
\name{kor_inspelningsgadget}
\alias{kor_inspelningsgadget}
\title{Kör inspelningsgadgeten}
\usage{
kor_inspelningsgadget(
  url = NULL,
  nedladdningsmapp = "C:/temp/nedladdningar",
  ...
)
}
\arguments{
\item{url}{Valfri adress att öppna direkt vid start.}

\item{nedladdningsmapp}{Läggs in i det genererade skriptets huvud som
\code{nedladdningsmapp <- ...}, om minst en nedladdning spelas in.}

\item{...}{Vidarebefordras till starta_skrapsession() (t.ex. browser_path).}
}
\description{
Startar en synlig skrapsession, injicerar inspelnings-JS:en, och visar
en liten kontrollpanel (miniUI-gadget) med en pollningstimer som var
500:e ms läser av nya händelser och uppdaterar en live-logg samt det
genererade skriptet.
}
