% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspelning.R
\docType{package}
\name{skrapa-package}
\alias{skrapa}
\alias{skrapa-package}
\title{Spela in klick/formulärval i en skrapsession och generera R-kod}
\description{
Bygger vidare på funktionerna i skrapning.R (samma paket). Tanken:
\enumerate{
\item injicera_inspelning() lägger en global JS-lyssnare i sidan som
fångar klick och select/input-ändringar, sparat i sessionStorage.
\item lasa_av_inspelning() läser av (och tömmer) loggen från R.
\item generera_skript() översätter de inspelade händelserna till anrop
mot klicka_via_id() / klicka_via_klass_och_text() / klicka_via_text()
/ selenider::elem_select() / kor_js()-fallback.
\item kor_inspelningsgadget() är ett litet Shiny-gadget (miniUI) som körs
LOKALT (inte deployat) och visar en enkel kontrollpanel bredvid det
synliga Edge-fönstret: Starta, Stoppa, Visa/kopiera genererad kod.
}
}
\details{
Utvecklarverktyg - shiny/miniUI/dplyr/tibble/purrr ligger i Suggests,
inte Imports, eftersom vanliga skrapskript (som bara använder
funktionerna i skrapning.R) inte ska behöva dra in dem.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FaluPeppe/skrapa}
  \item Report bugs at \url{https://github.com/FaluPeppe/skrapa/issues}
}

}
\author{
\strong{Maintainer}: Peter Efternamn \email{peter@example.se}

Authors:
\itemize{
  \item Peter Efternamn \email{peter@example.se}
}

}
\keyword{internal}
